# server/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart100X API",
...
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://smart100x-w9a4.vercel.app",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:5001",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/stats": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the Stats resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "General"
        ],
        "summary": "Retrieve Stats"
      }
    },
...
    "/api/youtube/get-upload-url": {
      "post": {
        "description": "This endpoint allows you to create data for the Youtube Get-upload-url resource. Authentication may be required depending on the route.",
        "parameters": [
          {
            "name": "origin",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileSize": {
                    "example": "any"
                  },
                  "contentType": {
                    "example": "any"
                  },
                  "title": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "YouTube"
        ],
        "summary": "Create Youtube Get-upload-url"
      }
    },
...
    "/api/oauth/youtube": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the Oauth Youtube resource. Authentication may be required depending on the route.",
        "parameters": [
          {
            "name": "onboarding",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "summary": "Retrieve Oauth Youtube"
      }
    },
    "/api/oauth/youtube/callback": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the Oauth Youtube Callback resource. Authentication may be required depending on the route.",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "summary": "Retrieve Oauth Youtube Callback"
      }
    },
...
    "/api/youtube/callback": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the Youtube Callback resource. Authentication may be required depending on the route.",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "YouTube"
        ],
        "summary": "Retrieve Youtube Callback"
      }
    },
    "/api/youtube/stats": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the Youtube Stats resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "YouTube"
        ],
        "summary": "Retrieve Youtube Stats"
      }
    },
    "/api/youtube/videos": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the Youtube Videos resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "YouTube"
        ],
        "summary": "Retrieve Youtube Videos"
      }
    },
    "/api/youtube/generate-thumbnail": {
      "post": {
        "description": "This endpoint allows you to create data for the Youtube Generate-thumbnail resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "YouTube"
        ],
        "summary": "Create Youtube Generate-thumbnail"
      }
    },
    "/api/youtube/generate-metadata": {
      "post": {
        "description": "This endpoint allows you to create data for the Youtube Generate-metadata resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "example": "any"
                  },
                  "options": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "YouTube"
        ],
        "summary": "Create Youtube Generate-metadata"
      }
    },
    "/api/youtube/access-token": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the Youtube Access-token resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "YouTube"
        ],
        "summary": "Retrieve Youtube Access-token"
      }
    },
...
    "/youtube": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the  Youtube resource. Authentication may be required depending on the route.",
        "parameters": [
          {
            "name": "onboarding",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "YouTube"
        ],
        "summary": "Retrieve  Youtube"
      }
    },
    "/youtube/callback": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the  Youtube Callback resource. Authentication may be required depending on the route.",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "YouTube"
        ],
        "summary": "Retrieve  Youtube Callback"
      }
    },
...
    "/callback": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the  Callback resource. Authentication may be required depending on the route.",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "tags": [
          "General"
        ],
        "summary": "Retrieve  Callback"
      }
    },
    "/stats": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the  Stats resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "General"
        ],
        "summary": "Retrieve  Stats"
      }
    },
    "/videos": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the  Videos resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "General"
        ],
        "summary": "Retrieve  Videos"
      }
    },
    "/generate-thumbnail": {
      "post": {
        "description": "This endpoint allows you to create data for the  Generate-thumbnail resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "General"
        ],
        "summary": "Create  Generate-thumbnail"
      }
    },
    "/generate-metadata": {
      "post": {
        "description": "This endpoint allows you to create data for the  Generate-metadata resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "example": "any"
                  },
                  "options": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "General"
        ],
        "summary": "Create  Generate-metadata"
      }
    },
    "/access-token": {
      "get": {
        "description": "This endpoint allows you to retrieve data for the  Access-token resource. Authentication may be required depending on the route.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "General"
        ],
        "summary": "Retrieve  Access-token"
      }
    },
    "/get-upload-url": {
      "post": {
        "description": "This endpoint allows you to create data for the  Get-upload-url resource. Authentication may be required depending on the route.",
        "parameters": [
          {
            "name": "origin",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fileSize": {
                    "example": "any"
                  },
                  "contentType": {
                    "example": "any"
                  },
                  "title": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "General"
        ],
        "summary": "Create  Get-upload-url"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}